requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
sqlmodel==0.0.21
SQLAlchemy==2.0.36
tenacity==9.0.0
openai==1.51.2
requests==2.32.3
urllib3==2.2.2
APScheduler==3.10.4
psycopg2-binary==2.9.9
